Kadane's Algorithm
Kadane’s algorithm is one of the famous approaches to solve the problem using dynamic programming
The time complexity of kadane’s algorithm for an array containing n integer element is O(n) as only one for loop is to be executed throughout the program. Similarly, the auxiliary space complexity of the algorithm is O(1).